Refactor: `ManagedPoolFactory` --> `ControlledManagedPoolFactory`; `BaseManagedPoolFactory` --> `ManagedPoolFactory`.
